Enable proper Painting and Fish listings, move old Paintings to UI.

+++ xivModdingFramework/Items/DataContainers/XivMinion.cs
@@ -63,7 +63,7 @@ namespace xivModdingFramework.Items.DataContainers
         /// </remarks>
         public XivDataFile DataFile { get; set; } = XivDataFile._04_Chara;
 
-        public ushort IconId { get; set; }
+        public uint IconId { get; set; }
 
         /// <summary>
         /// The Primary Model Information of the Minion Item
